Penyesuaian Module 2: Migration dan Model

+++ database/migrations/2025_12_20_070921_create_detail_orders_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -16,7 +15,7 @@ return new class extends Migration
             $table->foreignId('order_id')->constrained()->cascadeOnDelete();
             $table->foreignId('ticket_id')->constrained()->cascadeOnDelete();
             $table->integer('jumlah');
-            $table->integer('subtotal');
+            $table->decimal('subtotal', 10, 2);
             $table->timestamps();
         });
     }
